# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Agri System API Contract
  version: 1.0.0
  description: |
    Frontend integration contract generated from current Spring Boot + MyBatis-Plus controllers/services.

    Unified response wrapper:
    R<T> = { code:number, msg:string, data:T }
    - business success: code = 200
    - business failure: code != 200 (common: 500)

    Pagination response shape:
    data = { current:number, size:number, total:number, records:T[] }

    Authentication:
    Authorization: Bearer <token>
servers:
  - url: http://localhost:8080
...
paths:
...
  /system/user/list:
    get:
      tags: [SystemUser]
      summary: List users (paged)
      operationId: listUsers
      parameters:
        - in: query
          name: pageNum
          schema:
            type: integer
            default: 1
          example: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 10
          example: 10
        - in: query
          name: username
          schema:
            type: string
          example: farmer
        - in: query
          name: realName
          schema:
            type: string
          example: Li
        - in: query
          name: status
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: R<Page<SysUser>>
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PageSysUser'
              example:
                code: 200
                msg: Operation success
                data:
                  current: 1
                  size: 10
                  total: 1
                  records:
                    - userId: 3
                      username: farmer1
                      password: "123456"
                      realName: Farmer One
                      phone: "13800000003"
                      deptName: Planting Team
                      status: 1
                      createTime: "2026-02-28T09:00:00"
        '500':
          $ref: '#/components/responses/BizFail'
  /system/user/{userId}:
    get:
      tags: [SystemUser]
      summary: Get user detail
      operationId: getUserDetail
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
            format: int64
          example: 3
      responses:
        '200':
          description: R<SysUser>
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SysUser'
              example:
                code: 200
                msg: Operation success
                data:
                  userId: 3
                  username: farmer1
                  password: "123456"
                  realName: Farmer One
                  phone: "13800000003"
                  deptName: Planting Team
                  status: 1
                  createTime: "2026-02-28T09:00:00"
        '500':
          $ref: '#/components/responses/BizFail'
    delete:
      tags: [SystemUser]
      summary: Delete users by CSV ids
      operationId: deleteUsers
      parameters:
        - in: path
          name: userIds
          required: true
          schema:
            type: string
          description: Comma-separated ids, e.g. 3,4
          example: 3,4
      responses:
        '200':
          description: R<Void>
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RVoid'
              example:
                code: 200
                msg: Operation success
                data: null
        '500':
          $ref: '#/components/responses/BizFail'
  /system/user:
    post:
      tags: [SystemUser]
      summary: Create user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SysUser'
            example:
              username: farmer_new
              password: "123456"
              realName: New Farmer
              phone: "13800001111"
              deptName: Farm A
              status: 1
      responses:
        '200':
          description: R<Void>
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RVoid'
              example:
                code: 200
                msg: Operation success
                data: null
        '500':
          $ref: '#/components/responses/BizFail'
    put:
      tags: [SystemUser]
      summary: Update user
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SysUser'
            example:
              userId: 3
              realName: Farmer One Updated
              phone: "13800003333"
              deptName: Planting Team
              status: 1
      responses:
        '200':
          description: R<Void>
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RVoid'
              example:
                code: 200
                msg: Operation success
                data: null
        '500':
          $ref: '#/components/responses/BizFail'
  /system/role/list:
    get:
      tags: [SystemRole]
      summary: List roles (paged)
      operationId: listRoles
      parameters:
        - in: query
          name: pageNum
          schema:
            type: integer
            default: 1
          example: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 10
          example: 10
        - in: query
          name: roleName
          schema:
            type: string
          example: FARM
      responses:
        '200':
          description: R<Page<SysRole>>
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PageSysRole'
              example:
                code: 200
                msg: Operation success
                data:
                  current: 1
                  size: 10
                  total: 1
                  records:
                    - roleId: 3
                      roleName: Farmer
                      roleKey: FARMER
                      createTime: "2026-02-28T09:00:00"
        '500':
          $ref: '#/components/responses/BizFail'
  /system/role:
    post:
      tags: [SystemRole]
      summary: Create role
      operationId: createRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SysRole'
            example:
              roleName: Agronomist
              roleKey: AGRONOMIST
      responses:
        '200':
          description: R<Void>
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RVoid'
              example:
                code: 200
                msg: Operation success
                data: null
        '500':
          $ref: '#/components/responses/BizFail'
    put:
      tags: [SystemRole]
      summary: Update role
      operationId: updateRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SysRole'
            example:
              roleId: 6
              roleName: Demo Role
              roleKey: DEMO
      responses:
        '200':
          description: R<Void>
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RVoid'
              example:
                code: 200
                msg: Operation success
                data: null
        '500':
          $ref: '#/components/responses/BizFail'
  /system/role/{roleIds}:
    delete:
      tags: [SystemRole]
      summary: Delete roles by CSV ids
      description: |
        Role deletion rule:
        sys_user_role is many-to-many. Before deleting a role, bound user count must be checked.
        If bound count > 0, deletion is forbidden and returns code != 200.
      operationId: deleteRole
      parameters:
        - in: path
          name: roleIds
          required: true
          schema:
            type: string
          description: Comma-separated role ids, e.g. 3,4
          example: 6
      responses:
        '200':
          description: R<Void> (success or business-fail in body code)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RVoid'
              examples:
                success:
                  value:
                    code: 200
                    msg: Operation success
                    data: null
                boundFail:
                  value:
                    code: 500
                    msg: Role is bound by users, unbind first
                    data: null
        '500':
          $ref: '#/components/responses/BizFail'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BizFail:
      description: Business failure (R<Void>, code != 200)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RVoid'
          example:
            code: 500
            msg: Operation failed
            data: null
...
  schemas:
    R:
      type: object
      properties:
        code:
          type: integer
          description: Business code, 200 success, non-200 fail
        msg:
          type: string
        data:
          nullable: true
      required: [code, msg, data]
    RVoid:
      allOf:
        - $ref: '#/components/schemas/R'
        - type: object
          properties:
            data:
              nullable: true
              example: null
    PageBase:
      type: object
      properties:
        current:
          type: integer
          format: int64
        size:
          type: integer
          format: int64
        total:
          type: integer
          format: int64
      required: [current, size, total]
    PageSysUser:
      allOf:
        - $ref: '#/components/schemas/PageBase'
        - type: object
          properties:
            records:
              type: array
              items:
                $ref: '#/components/schemas/SysUser'
          required: [records]
    PageSysRole:
      allOf:
        - $ref: '#/components/schemas/PageBase'
        - type: object
          properties:
            records:
              type: array
              items:
                $ref: '#/components/schemas/SysRole'
          required: [records]
...
    SysUser:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        username:
          type: string
        password:
          type: string
        realName:
          type: string
        phone:
          type: string
        deptName:
          type: string
        status:
          type: integer
          description: 1 enabled, 0 disabled
        createTime:
          type: string
          format: date-time
    SysRole:
      type: object
      properties:
        roleId:
          type: integer
          format: int64
        roleName:
          type: string
        roleKey:
          type: string
        createTime:
          type: string
          format: date-time
...